Update User Management Form

+++ Forms/frmLoai.cs
@@ -75,16 +75,19 @@ namespace Forms
 
         private void btnXoa_Click(object sender, EventArgs e)
         {
-            try
+            if(MessageBox.Show("Bạn có chắc chắn muốn xoá thể loại này không? Lưu ý các sản phẩm hiện có thuộc thể loại này sẽ bị xoá?", "Cảnh Báo", MessageBoxButtons.OKCancel) == DialogResult.OK)
             {
-                int CategoryId = int.Parse(txtMaLoai.Text);
-                productCategoryService.DeleteCategory(CategoryId);
-                List<ProductCategory> categories = productCategoryService.GetAllCategories();
-                FillDGV(categories);
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
+                try
+                {
+                    int CategoryId = int.Parse(txtMaLoai.Text);
+                    productCategoryService.DeleteCategory(CategoryId);
+                    List<ProductCategory> categories = productCategoryService.GetAllCategories();
+                    FillDGV(categories);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
             }
         }
